+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Minor refactors.
 
+### Added
+
+- `ServiceClusterOptions` and optional (experimental) `autoRenew` option.
+
 ## [0.4.1] - 2023-5-2
 
 ### Changed